requirements.txt
click==8.1.3
python3-discogs-client
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.9.0
pymongo~=4.2.0
multipledispatch~=0.6.0
networkx~=2.8.7
bokeh~=2.4.3
scipy==1.9.3
